bataille: new example card game
Bataille is a french card game that is very easy to play
2 opponents, the same number of cards. Each opponent plays
the card that is on top of his pile, the card that is ranked
higher than the other win. If the cards have same rank, it's
"Bataille", you play again. Whenever you win, you get all the
cards back to the bottom of your pile.
The first player to win all the cards win.

Signed-off-by: Gilles Talis <gilles.talis@gmail.com>